Move action classes into different files and export from main module
